Refactor start time estimation, concurrency oracle, and resource availability to enhancement.
