+++ tests/pix_framework/enhancement/test_resource_availability.py
@@ -2,8 +2,8 @@ from pathlib import Path
 
 import pandas as pd
 from pix_framework.discovery.resource_calendar_and_performance.crisp.resource_calendar import Interval, RCalendar
-from pix_framework.discovery.start_time_estimator.config import Configuration
-from pix_framework.discovery.start_time_estimator.resource_availability import (
+from pix_framework.enhancement.start_time_estimator.config import Configuration
+from pix_framework.enhancement.resource_availability import (
     CalendarResourceAvailability,
     SimpleResourceAvailability,
 )
